Introduce succinct per-section leading text
As per discussions in #14, introduces a per-section leading text that's
meant to be as succinct as possible. This is eventually the only part of
the section that will be visible on the index, with all other content
moved to a separate page.

For now, it's kept as a single document to ease change while the content
matures.